Add CIFAR10 images for prediction and CNN prediction scripts
Add a set of images that can be used to run inference requests
against a CIFAR10 model.

Add the following scripts imported from 3rd party sources
and slightly modified to be more flexible to be used with any
CNN models served under an API that is compatible with the
TensorFlow serving REST API:

- img_to_npy.py (imported from [1]) - this script is used
to convert a set of raw images into a numpy dataset. The numpy
dataset format is used by other scripts as input to perform batch
prediction request or other similar functions

- generate_rest_payload.py - use this script to generate a JSON
request payload that can be used for TensorFlow REST prediction
requests (e.g. with curl). It takes in a numpy dataset and outputs
a JSON file with a CNN prediction request payload

- rest_binary_client.py (imported from [2]) - an OpenVINO Model
Server client utility that can be used to send REST binary [3] prediction
requests to an OpenVINO Model Server starting from an input numpy dataset.
The original script has been slightly modified to allow it to work
with any model, not just CIFAR-100.

- rest_serving_client.py (imported from [4]) - an OpenVINO Model
Server client utility that can be used to send REST prediction
requests to a TensorFlow serving API starting from an input numpy dataset.
The original script has been slightly modified to allow it to work
with any model, not just CIFAR-100. Note that this script works with
any inference service as long as its API is compatible with the
TensorFlow REST serving API. It should also work with any CNN model.

[1] https://github.com/foolmarks/images_to_npy/blob/master/img_to_npy.py
[2] https://github.com/openvinotoolkit/model_server/blob/main/example_client/rest_binary_client.py
[3] https://github.com/openvinotoolkit/model_server/blob/main/docs/binary_input.md
[4] https://github.com/openvinotoolkit/model_server/blob/main/example_client/rest_serving_client.py